requirements.txt
pandas
pytesseract
PyMuPDF
Pillow
reportlab
pdfplumber
